% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperChemoBridge-package.R
\docType{package}
\name{hyperChemoBridge-package}
\alias{hyperChemoBridge-package}
\alias{hyperChemoBridge}
\title{Bridge between hyperSpec & ChemoSpec packages}
\description{
Functions for converting between hyperSpec and ChemoSpec (Spectra) objects
as used by the respective packages.
}
\author{
Conor McManus
}
\keyword{package}
