% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra2hyperSpec.R
\name{spectra2hyperSpec}
\alias{spectra2hyperSpec}
\title{Convert a Spectra object to a hyperSpec object}
\usage{
spectra2hyperSpec(Spectra)
}
\arguments{
\item{Spectra}{A \code{\link[ChemoSpec]{Spectra}} object.}
}
\value{
A \code{\link[=setClass]{hyperspec}} object.
}
\description{
This function converts a \code{Spectra} object as used the in the \code{ChemoSpec}
package to a \code{hyperSpec} object as used in the package of the same name.
}
\examples{
require("hyperSpec")
require("ChemoSpec")
data(SrE.IR) # data set from ChemoSpec
IR <- spectra2hyperSpec(SrE.IR)
IR
str(IR)

}
\author{
Conor McManus
}
\keyword{utilities}
