\name{hyperChemoBridge-package}
\alias{hyperChemoBridge-package}
\alias{hyperChemoBridge}
\docType{package}
\title{
Bridge between hyperSpec & ChemoSpec packages
}
\description{
Functions for converting between hyperSpec and ChemoSpec (Spectra) objects as used by the respective packages.
}
\author{
Conor McManus.\cr
Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\references{
\url{https://github.com/Chathurga/HyperChemoBridge}
}
\keyword{ package }
