\name{hyperChemoBridge-package}
\alias{hyperChemoBridge-package}
\alias{hyperChemoBridge}
\docType{package}
\title{
Bridge between hyperSpec & ChemoSpec
}
\description{
Functions for converting between hyperSpec and ChemoSpec (Spectra) objects
}
\details{
\tabular{ll}{
Package: \tab hyperChemoBridge\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2014-01-23\cr
License: \tab GPL-3\cr
}
}
\author{
Conor McManus.\cr
Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\references{
\url{https://github.com/Chathurga/HyperChemoBridge}
}
\keyword{ package }
