\name{HyperChemoBridge-package}
\alias{HyperChemoBridge-package}
\alias{HyperChemoBridge}
\docType{package}
\title{
HyperSpec/ChemoSpec Bridge
}
\description{
Methods for converting between HyperSpec and ChemoSpec objects
}
\details{
\tabular{ll}{
Package: \tab HyperChemoBridge\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-27\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Conor McManus
Maintainer: ? <yourfault@somewhere.net>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
